requirements.txt
google-cloud-bigquery==3.40.0
google-cloud-aiplatform==1.136.0
google-genai==1.62.0
google-cloud-logging==3.13.0
tenacity==9.1.2
fastapi==0.128.1
uvicorn==0.40.0
pydantic==2.12.5
pytest==9.0.2
langgraph==0.3.5
PyPDF2==3.0.1
